requirements.txt
gitdb==4.0.12
GitPython==3.1.44
smmap==5.0.2
watchdog==6.0.0
